loader/requirements.txt
google-cloud-bigquery==3.25.0
PyMySQL==1.1.1
